Rename package testSuiteClass to testsuite
