+++ jwala-integration-test/src/test/java/com/cerner/jwala/ui/selenium/testsuite/PreFlightTest.java
@@ -1,4 +1,4 @@
-package com.cerner.jwala.ui.selenium.testSuiteClasses;
+package com.cerner.jwala.ui.selenium.testsuite;
 
 import org.junit.Test;
