+++ jwala-integration-test/src/test/java/com/cerner/jwala/ui/selenium/testsuite/configuration/resources/ResourceTopologyTest.java
@@ -1,7 +1,7 @@
-package com.cerner.jwala.ui.selenium.testSuiteClasses.configuration.resources;
+package com.cerner.jwala.ui.selenium.testsuite.configuration.resources;
 
 import com.cerner.jwala.ui.selenium.SeleniumTestHelper;
-import com.cerner.jwala.ui.selenium.testSuiteClasses.JwalaTest;
+import com.cerner.jwala.ui.selenium.testsuite.JwalaTest;
 import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
